% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colmozzie}
\alias{colmozzie}
\title{Weekly notified dengue cases in Sri Lanka}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 279 rows and 12 columns.}
\source{
weekly epidemiological reports published by the Epidemiological Unit, Ministry of Health, Sri Lanka
}
\usage{
colmozzie
}
\description{
Weekly notified dengue cases and climate variables in Colombo districts of Sri Lanka from 2008/week-52 to 2014/week-21
}
\examples{
data(colmozzie)
head(colmozzie)
summary(colmozzie)
}
\keyword{datasets}
